+++ src/dashboardWebView/components/Header/Navigation.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { useRecoilState, useRecoilValue } from 'recoil';
 import { Tab } from '../../constants/Tab';
-import { AllPagesAtom, SettingsAtom, TabAtom, TabInfoAtom } from '../../state';
+import { SettingsAtom, TabAtom, TabInfoAtom } from '../../state';
 import * as l10n from '@vscode/l10n';
 import { LocalizationKey } from '../../../localization';
 
@@ -40,24 +40,22 @@ const NavigationItem: React.FunctionComponent<INavigationItemProps> = ({
 export const Navigation: React.FunctionComponent<INavigationProps> = ({
 
 }: React.PropsWithChildren<INavigationProps>) => {
-  const pages = useRecoilValue(AllPagesAtom);
   const [crntTab, setCrntTab] = useRecoilState(TabAtom);
   const tabInfo = useRecoilValue(TabInfoAtom);
   const settings = useRecoilValue(SettingsAtom);
 
-  const tabs = [
-    { name: l10n.t(LocalizationKey.dashboardHeaderNavigationAllArticles), id: Tab.All },
-    { name: l10n.t(LocalizationKey.dashboardHeaderNavigationPublished), id: Tab.Published },
-    { name: l10n.t(LocalizationKey.dashboardHeaderNavigationDraft), id: Tab.Draft }
-  ];
+  const tabs = React.useMemo(() => {
+    const crntTabs = [
+      { name: l10n.t(LocalizationKey.dashboardHeaderNavigationAllArticles), id: Tab.All }
+    ];
 
-  const usesDraft = React.useMemo(() => {
-    return pages.some((x) => x.fmDraft);
-  }, [pages]);
+    if (settings?.draftField?.type === 'boolean' && tabInfo && Object.keys(tabInfo).length > 1) {
+      crntTabs.push({ name: l10n.t(LocalizationKey.dashboardHeaderNavigationPublished), id: Tab.Published });
+      crntTabs.push({ name: l10n.t(LocalizationKey.dashboardHeaderNavigationDraft), id: Tab.Draft });
+    }
 
-  if (!usesDraft) {
-    return null;
-  }
+    return crntTabs;
+  }, [settings?.draftField?.type, tabInfo]);
 
   return (
     <nav className="flex-1 -mb-px flex space-x-6 xl:space-x-8" aria-label="Tabs">